requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers>=2.7.0
weaviate-client==4.4.4
